requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart


##-e .